Refactor(code-quality): A single function to handle payment
