+++ imports/plugins/core/layout/client/index.js
@@ -17,6 +17,7 @@ import "./templates/layout/notFound/notFound.html";
 import "./templates/layout/notFound/notFound.js";
 import "./templates/layout/notice/unauthorized.html";
 import "./templates/layout/layout.html";
+import "./templates/layout/wallet/payment.js";
 import "./templates/layout/wallet/wallet.html";
 import "./templates/layout/wallet/wallet.js";
