+++ imports/plugins/core/layout/client/templates/layout/wallet/wallet.js
@@ -1,7 +1,6 @@
 import { Meteor } from "meteor/meteor";
 import { Template } from "meteor/templating";
-import { Reaction } from "/client/api";
-import { Wallet, Accounts, Packages, Shops } from "/lib/collections";
+import { Wallet, Accounts } from "/lib/collections";
 
 Template.wallet.events({
   "submit #deposit": (event) => {
@@ -24,12 +23,12 @@ Template.wallet.events({
       Alerts.toast("Insufficient Balance", "error");
       return false;
     }
-    const to = document.getElementById("recipient").value;
+    const recipient = document.getElementById("recipient").value;
     amount /= getExchangeRate();
-    const transaction = { amount, to, date: new Date(), transactionType: "Debit" };
+    const transaction = { amount, to: recipient, date: new Date(), transactionType: "Debit" };
     Meteor.call("wallet/transaction", Meteor.userId(), transaction, (err, res) => {
       if (res === 2) {
-        Alerts.toast(`No user with email ${to}`, "error");
+        Alerts.toast(`No user with email ${recipient}`, "error");
       } else if (res === 1) {
         document.getElementById("recipient").value = "";
         document.getElementById("transferAmount").value = "";
@@ -62,55 +61,6 @@ Template.wallet.helpers({
     return Template.instance().state.get("details").transactions;
   }
 });
-const getPaystackConfig = () => {
-  return Packages.findOne({
-    name: "paystack",
-    shopId: Reaction.getShopId()
-  });
-};
-
-const handlePayment = (transactionId) => {
-  const paystackConfig = getPaystackConfig();
-  HTTP.call("GET", `https://api.paystack.co/transaction/verify/${transactionId}`,
-    { headers: { Authorization: `Bearer ${paystackConfig.settings.secretkey}`}},
-    function (error, response) {
-      if (error) {
-        Alerts.toast("Unable to verify payment", "error");
-      } else if (response.data.data.status !== "success") {
-        Alerts.toast("Payment was unsuccessful", "error");
-      } else {
-        const exchangeRate =  getExchangeRate();
-        const paystackResponse = response.data.data;
-        paystackMethod = {
-          processor: "Paystack",
-          storedCard: paystackResponse.authorization.last4,
-          method: "Paystack",
-          transactionId: paystackResponse.reference,
-          currency: paystackResponse.currency,
-          amount: paystackResponse.amount,
-          status: paystackResponse.status,
-          mode: "authorize",
-          createdAt: new Date(),
-          transactions: { amount: paystackResponse.amount / (100 * exchangeRate), referenceId: paystackResponse.reference,
-            date: new Date(), transactionType: "Credit" }
-        };
-        transactions.amount /= exchangeRate;
-        Meteor.call("wallet/transaction", Meteor.userId(), paystackMethod.transactions, (err, res) => {
-          if (res) {
-            document.getElementById("depositAmount").value = "";
-            Alerts.toast("Your deposit was successful", "success");
-          } else {
-            Alerts.toast("An error occured, please try again", "error");
-          }
-        });
-      }
-    });
-};
-
-getExchangeRate = () => {
-  const shop = Shops.find(Reaction.getShopId()).fetch();
-  return shop[0].currencies.NGN.rate;
-};
 
 // Paystack payment
 payWithPaystack = (email, amount) => {
@@ -120,7 +70,7 @@ payWithPaystack = (email, amount) => {
     email: email,
     amount: amount * 100,
     callback: function (response) {
-      handlePayment(response.reference);
+      handlePayment(response.reference, "deposit");
     }
   });
   handler.openIframe();
